git-svn-id: svn+ssh://software-srn.sandia.gov/svn/spparks/trunk@1063 685d9ffe-9724-0410-857b-ce06367cce3e
Former-commit-id: 98b1b5d86ae0150b4970ff895a886da6a75b2759

+++ src/app_sinter.h
@@ -106,8 +106,8 @@ class AppSinter : public AppLattice {
   void stats(char *strtmp);
 
   // Checking lattice configurations
-  long double count_vacant();
-  long double count_grain_sites();
+  double count_vacant();
+  double count_grain_sites();
   double calculate_density();
