+++ src/diag_sinter_free_energy.cpp
@@ -56,10 +56,10 @@ void DiagSinterFreeEnergy::compute()
   
   const int VACANT ( AppSinter::VACANT );
 
-  long double interfacialFEtmp = 0.0;
+  double interfacialFEtmp = 0.0;
   for (int i = 0; i < nlocal; i++) {
 	int ispin = spin[i];
-	long double surface = 0;
+	double surface = 0;
 	if ( ispin > VACANT ) { // If I am a grain site add the number of neighbors that are pore sites
 		for (int j = 0; j < numneigh[i]; j++)
 			if (spin[neighbor[i][j]] == VACANT) surface++;
@@ -67,7 +67,7 @@ void DiagSinterFreeEnergy::compute()
 	interfacialFEtmp += surface;
 	//interfacialFEtmp += appsinter->site_surface(i);
   }
-  MPI_Allreduce(&interfacialFEtmp,&interfacialFE,1,MPI_LONG_DOUBLE,MPI_SUM,world);
+  MPI_Allreduce(&interfacialFEtmp,&interfacialFE,1,MPI_DOUBLE,MPI_SUM,world);
 //  density = appsinter->calculate_density();	
 }
